Add `PaymentSheetScaffold` for screen scaffolding
