+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/ui/PaymentOptionsScreen.kt
@@ -1,24 +1,17 @@
 package com.stripe.android.paymentsheet.ui
 
-import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidViewBinding
-import androidx.compose.ui.zIndex
 import com.stripe.android.link.ui.verification.LinkVerificationDialog
 import com.stripe.android.paymentsheet.PaymentOptionsViewModel
 import com.stripe.android.paymentsheet.R
@@ -34,32 +27,13 @@ internal fun PaymentOptionsScreen(
     viewModel: PaymentOptionsViewModel,
     modifier: Modifier = Modifier,
 ) {
-    val scrollState = rememberScrollState()
-
-    val targetElevation by remember {
-        derivedStateOf {
-            if (scrollState.value > 0) {
-                8.dp
-            } else {
-                0.dp
-            }
-        }
-    }
-
-    val elevation by animateDpAsState(targetValue = targetElevation)
-
-    Column {
-        // We need to set a z-index to make sure that the Surface's elevation shadow is rendered
-        // correctly above the screen content.
-        Surface(elevation = elevation, modifier = Modifier.zIndex(1f)) {
-            PaymentSheetTopBar(viewModel)
-        }
-
-        PaymentOptionsScreenContent(
-            viewModel = viewModel,
-            modifier = modifier.verticalScroll(scrollState),
-        )
-    }
+    PaymentSheetScaffold(
+        topBar = { PaymentSheetTopBar(viewModel) },
+        content = { scrollModifier ->
+            PaymentOptionsScreenContent(viewModel, scrollModifier)
+        },
+        modifier = modifier,
+    )
 }
 
 @Composable
